[Model] Explicitly `default_pooling_type` interface
